==> Data_Preparation/1 - Trailers Viewed and Aggregate Table, by Algorithm.ipynb <==
Data Date Range:
    <br />Start time: 2016-05-05 09:40:26
    <br />End time: 2018-01-17 01:59:42

#### This notebook:
- Generate a table with total trailer viewed and their algorithm assignment
- Generate a table with recommendations that are clicked and their algorithm assignment
- Generate a aggregate table with count_total_played, count_rec_played, and algorithm assignment

#### We will be using the 'log_trailer_actions_sessionIds' file, and here are some explanations of the 'actions' in this file. ####       
##### If a user opens the trailer from the movie detail page: ####
* TrailerModalLaunched: If a user opens the trailer from the movie detail page
* PlayedFromLaunch: Happens right after 'TrailerModalLaunched'

##### User clicked in the trailer interface to go to a different trailer recommended: ####
* PlayedFromNext
* PalyedFromPrevious
* PlayedFromRecommendation


#### This notebook generates the following table:
- rec_played_alg: table with after recommendation clicks and algorithm assignment
- total_rec_refresh_aggregate: an aggretable with count_total_played, count_rec_played, and algorithm assignment


### Import ###

import pandas as pd
import numpy as np

#### Need to run 0 - Generate log_trailer_actions Dataframe before doing the following

action_data = pd.read_csv('../Clean_Data/log_trailer_actions_sessionIds.csv')

### Load the dataset with algorithm assignment in each session in the experiment

# Get the algorithm assignment
expt_trailer = pd.read_csv('../Clean_Data/expt_trailer.csv', sep=',')
alg_assignment = expt_trailer[['userId','loginId','recommenderId']]
alg_assignment.sort_values(by='recommenderId').shape
alg_assignment.head()

### Rec_played (recommended trailers that are clicked on) and Algorithm

If an action is among: "PlayedFromNext", "PlayedFromRecommendation", "PlayedFromPrevious", it means a recommendation is played, so we keep this row in the table.

# Trailers recommended by an algorithm and played, excluding 'PlayedFromYoutubePlayer'
rec_played = action_data[action_data['action'].isin(["PlayedFromNext", "PlayedFromRecommendation", "PlayedFromPrevious"])]

# Join the table of trailers recommended and played with the table of algorithm assignment
rec_played_alg = pd.merge(rec_played, alg_assignment, on=['loginId','userId'], how='left')
rec_played_alg.head()

# Drop rows without recommender assignment
rec_played_alg2 = rec_played_alg.dropna(subset=['recommenderId'])


rec_played_alg2.count()

rec_played_alg2.to_csv("../Clean_Data/rec_played_alg.csv", index=False)

# Find out the count of recommended trailers played in a session:
rec_played_alg_count = rec_played_alg2.groupby(['userId','loginId','recommenderId'])['movieId'].count().reset_index(name="count_rec_played")
rec_played_alg_count.head()

#### Get content information of these recommendations

rec_seed_info = pd.read_csv("../Clean_Data/rec_seed_info.csv")

rec_seed_info.head()

rec_played_info = pd.merge(rec_played_alg2, rec_seed_info, on=['loginId','movieId'],how='left')
# Drop NA
rec_played_info.dropna(inplace=True)


rec_played_info.drop_duplicates(subset=['loginId','movieId','SeedMovie']).count()

### Total_played (All trailers that are clicked on) and Algorithm

If an action is among: "PlayedFromLaunch","PlayedFromNext", "PlayedFromRecommendation", "PlayedFromPrevious", it means a trailer is played, so we keep this row in the table.

# All trailers played, with their algorithm assignment. 
total_played = action_data[action_data['action'].isin(["PlayedFromLaunch","PlayedFromNext", "PlayedFromRecommendation", "PlayedFromPrevious"])].copy()
total_played.count()

total_played.drop("Unnamed: 0",axis=1, inplace=True)

# Merge total_played with algorithm assignment
total_played_alg = pd.merge(total_played, alg_assignment, on=['loginId','userId'], how='left')

# Drop na values
total_played_alg2 = total_played_alg.dropna(subset=['recommenderId']).copy()

total_played_alg2.count()

total_played_alg2.to_csv("../Clean_Data/total_played_alg.csv", index=False)

### Aggretable count_total_played, count_rec_played, refreshes, with algorithm assignment, by loginId

# Find out the count of total trailers played in a session:
total_played_alg_count = total_played_alg2.groupby(['userId','loginId','recommenderId'])['movieId'].count().reset_index(name="count_total_played")

# Join the two tables above:
total_rec_count = pd.merge(total_played_alg_count, rec_played_alg_count, on=['loginId','userId'], how='left')


rec_refresh = action_data[action_data['action'] == 'RecommendationsRefreshed']
refresh_alg = pd.merge(rec_refresh, alg_assignment, on=['loginId','userId'], how='left')

refresh_alg_count = refresh_alg.groupby(['userId','loginId','recommenderId'])['movieId'].count().reset_index(name="count_refresh")
refresh_alg_count.head()

# Join the table of #refresh with the previous aggregate table
total_rec_refresh_count = pd.merge(total_rec_count, refresh_alg_count, on=['loginId','userId'], how='left')

## Drop NA
total_rec_refresh_count['count_refresh'].fillna(0, inplace=True)
total_rec_refresh_count['count_rec_played'].fillna(0, inplace=True)


# Rename the column: recommenderId_x as Algorithm
total_rec_refresh_count=total_rec_refresh_count.rename(columns = {'recommenderId_x':'Algorithm'})

# Drop the columns recommenderId_y and recommenderId
total_rec_refresh_count.drop('recommenderId_y', axis = 1, inplace = True)
total_rec_refresh_count.drop('recommenderId', axis = 1, inplace = True)

total_rec_refresh_count.head()

Now, we have generated an aggregate table with total played, rec played, and #refresh in each login session. 

# Write result to a csv file
total_rec_refresh_count.to_csv("../Clean_Data/total_rec_refresh_aggregate.csv")
