==> Data_Preparation/2 - Action counts (Like, Dislike, Wishlist, DetailModel, Watch>50%) - by Algorithm.ipynb <==
Data Date Range:
    <br />Start time: 2016-05-05 09:40:26
    <br />End time: 2018-01-17 01:59:42

#### This notebook generates the following table:
- actions_alg: a table of key interactions, and corresponding information about the user and the movie
- interactions_rec_ratio: the ratio of key interactions over rec clicks, broken down by algorithms

#### Explanations of selected actions:
A list of the actions:

    ['TrailerModalLaunched', 'PlayedFromLaunch', 'PlayedFromNext',
    'PlayedFromRecommendation', 'MovieDetailsModalOpened',
    'TrailerPaused', 'MovieDetailsModalClosed', 'TrailerModalShutdown',
    'IssueRaised', 'PlayedFromYoutubePlayer', 'TrailerEnded',
    'TrailerLiked', 'GoToMovieDetails', 'RecommendationsRefreshed',
    'TrailerUnliked', 'TrailerDisliked', 'TrailerUndisliked',
    'WishlistedMovie', 'UnwishlistedMovie', 'PlayedFromPrevious',
    'RatedMovie']
       
##### If a user opens the trailer from the movie detail page: ####
* TrailerModalLaunched: If a user opens the trailer from the movie detail page
* PlayedFromLaunch: Happens right after 'TrailerModalLaunched'

##### User clicked in the trailer interface to go to a different trailer recommended: ####
* PlayedFromNext
* PalyedFromPrevious
* PlayedFromRecommendation


### Import and run code from another notebook

import pandas as pd
import numpy as np

%run '1 - Trailers Viewed and Aggregate Table, by Algorithm.ipynb'


### Generate the dataset with the counts of each action  ###

Except for the actions "PlayedFromLaunch","PlayedFromNext", "PlayedFromRecommendation", "PlayedFromPrevious" that indicate if a trailer is played or not, there are also other informative actions. Therefore, in this section, we count how many times each action happened. 

action_data = pd.read_csv('../Clean_Data/log_trailer_actions_sessionIds.csv')

# Get the algorithm assignment
expt_trailer = pd.read_csv('../Clean_Data/expt_trailer.csv', sep=',')
alg_assignment = expt_trailer[['userId','loginId','recommenderId']]

# generate the dataset with the counts of each action
actions_alg = pd.merge(action_data, alg_assignment, on=['loginId','userId'], how='left')
actions_alg.drop(['Unnamed: 0'],axis=1, inplace=True)
interactions_alg = actions_alg[actions_alg['action'].isin(['TrailerLiked','TrailerDisliked','MovieDetailsModalOpened','WishlistedMovie','GoToMovieDetails'])]

interactions_alg.groupby('action').count()

#### Find 'WatchedMoreThanHalf'

# Data ends on 2018-01-17
watch = pd.read_csv('../Clean_Data/log_trailer_watch.csv')
watch_small = watch[watch['tstamp']<"2018-01-17 01:59:42"]

mlusers = [32451,49002,49253,50540,51234,51714,65817,66644,68903,70023,70026,92721,92723,95195,96279,96483,96691,97397,104218,113699,114092,114844,122798,124450,124790,125283,125431,127628,127979,129299,130353,132131,132145,132637,132726,132728,132871,133047,134070,134088,134143,134235,134342,134349,134372,134373,134393,134484,134496,134498,134504,134533,135260,135670,136346,136505,136995,137096,137331,137372,138083,138089,138090,138108,141799,142115,143473,149815,151709,151982,166008,173637,176970,183038,183567,197881,202657,207221,221192,221515,225510,233659,233985,237287,237332,238576,238876,240312,242238,242683,243010,243445,245696,246018,246058,246587,246828,248583,251610,251922,252238,252534,253066,253866,253869,257628,257629,257756,258553,258554,258555,258556,258557,258558,258559,258560,258561,258562,258563,258564,258565,259056,260701,264178,266542,268253,269945,286014,287206,292397,301040,304542,313322,316070,317416,318334]

watch_noml = watch[~watch.userId.isin(mlusers)].copy()

watch_noml.sort_values('percentWatched',ascending=False, inplace=True)

watch_noml = watch_noml[watch_noml['percentWatched']<=100].copy()

watch_noml.sort_values(by=['loginId','movieId'],inplace=True)

# Keep only the largest percent watched.
watch_noml_small = watch_noml.drop_duplicates(subset=["loginId","movieId"], keep="first").copy()
# watch_noml_small.groupby('percentWatched').count()
watch_noml_small.head()

watch_more_than_half = watch_noml_small[watch_noml_small['percentWatched'] >= 50]
watch_more_than_half.count()

watch = pd.read_csv('../Clean_Data/log_trailer_watch.csv')

# Exclude movielens users
mlusers = [32451,49002,49253,50540,51234,51714,65817,66644,68903,70023,70026,92721,92723,95195,96279,96483,96691,97397,104218,113699,114092,114844,122798,124450,124790,125283,125431,127628,127979,129299,130353,132131,132145,132637,132726,132728,132871,133047,134070,134088,134143,134235,134342,134349,134372,134373,134393,134484,134496,134498,134504,134533,135260,135670,136346,136505,136995,137096,137331,137372,138083,138089,138090,138108,141799,142115,143473,149815,151709,151982,166008,173637,176970,183038,183567,197881,202657,207221,221192,221515,225510,233659,233985,237287,237332,238576,238876,240312,242238,242683,243010,243445,245696,246018,246058,246587,246828,248583,251610,251922,252238,252534,253066,253866,253869,257628,257629,257756,258553,258554,258555,258556,258557,258558,258559,258560,258561,258562,258563,258564,258565,259056,260701,264178,266542,268253,269945,286014,287206,292397,301040,304542,313322,316070,317416,318334]

watch_small = watch[watch['tstamp']<"2018-01-17 01:59:42"]
watch_noml = watch_small[~watch_small.userId.isin(mlusers)]
watch_noml = watch_noml.drop(['youtubeTrailerId'],axis=1).copy()

watch_noml_small = watch_noml[watch_noml['percentWatched']<=100].copy()
watch_noml_small.sort_values('percentWatched',ascending=False, inplace=True)
watch_noml_small.sort_values(by=['loginId','movieId'],inplace=True)

watch_nodup = watch_noml_small.drop_duplicates(subset=["loginId","movieId"], keep="first").copy()

# Keep only the highest percent watched for each trailer view
watch_nodup.groupby('percentWatched').count()

expt_trailer = pd.read_csv('../Clean_Data/expt_trailer.csv')
watch_alg = pd.merge(watch_nodup, expt_trailer, on=['loginId','userId'], how="inner") # Use inner to drop sessions without alg assignment

watch_more_than_half = watch_alg[watch_alg['percentWatched'] >= 50].copy()

watch_more_than_half.count()

watch_more_than_half_small = watch_more_than_half[['loginId','movieId','recommenderId','userId','positionInList','tstamp_x']].copy()
watch_more_than_half_small['action']='WatchedMoreThanHalf'
watch_more_than_half_small.rename(columns={"positionInList":"position",'tstamp_x':'tstamp'}, inplace=True)

interactions_alg = interactions_alg.append(watch_more_than_half_small)

interactions_alg.groupby("action").count()

#### Finish Appending watch_more_than_half to actions_alg  ⤴️

### Interactions with recommended trailers clicked

interactions_alg.count()

interactions_alg.to_csv('../Clean_Data/actions_alg.csv', sep=',')

actions_alg_rec = actions_alg[actions_alg['action'].isin(["PlayedFromNext", "PlayedFromRecommendation", "PlayedFromPrevious"])]
actions_alg_rec_small = actions_alg_rec[['loginId','movieId','action','tstamp','userId','recommenderId']]
actions_alg_rec_small.head()

# Actions taken place AFTER user clicked on the recommended trailer
interactions_alg_agg = pd.merge(interactions_alg, actions_alg_rec_small, on=['loginId','movieId'],how='inner')
interactions_alg_agg.count()

# Count the number of times an action taken place AFTER user clicked on a recommended trailer
interactions_alg_agg_small = interactions_alg_agg[interactions_alg_agg['tstamp_x']>interactions_alg_agg['tstamp_y']].copy()


interactions_alg_group = interactions_alg_agg_small.groupby(['action_x','recommenderId_x']).count().reset_index()[['action_x','recommenderId_x','loginId']]
interactions_alg_group.rename(columns = {"action_x":"action", "recommenderId_x":"recommenderId","loginId":"count"},inplace=True)

# Action_alg_rec_small is created abooove

actions_alg_rec_agg = actions_alg_rec_small.groupby('recommenderId').count().reset_index()
actions_alg_rec_agg = actions_alg_rec_agg[['recommenderId','loginId']].copy()
actions_alg_rec_agg.rename(columns={"loginId":"count_rec_clicks"},inplace=True)

actions_alg_rec_agg.head()

interactions_alg_rec_percent = pd.merge(interactions_alg_group, actions_alg_rec_agg, on="recommenderId", how="left")
interactions_alg_rec_percent['ratio_action_rec_clicks'] = interactions_alg_rec_percent['count']/interactions_alg_rec_percent['count_rec_clicks']

interactions_alg_rec_percent

interactions_alg_rec_percent.to_csv('../Clean_Data/interactions_rec_ratio_9_15.csv')


### Code above are run in September 2018. Because the data in the database changed, results from code above are slightly different.
### Codes below are written and run in May, 2018. Don't edit them. Leave them for comparison.

# Actions taken place AFTER user clicked on the recommended trailer
interactions_alg_agg = pd.merge(interactions_alg, actions_alg_rec_small, on=['loginId','movieId'],how='inner')
interactions_alg_agg.count()

# Count the number of times an action taken place AFTER user clicked on a recommended trailer
interactions_alg_agg_small = interactions_alg_agg[interactions_alg_agg['tstamp_x']>interactions_alg_agg['tstamp_y']].copy()


interactions_alg_group = interactions_alg_agg_small.groupby(['action_x','recommenderId_x']).count().reset_index()[['action_x','recommenderId_x','loginId']]
interactions_alg_group.rename(columns = {"action_x":"action", "recommenderId_x":"recommenderId","loginId":"count"},inplace=True)

# Action_alg_rec_small is created abooove

actions_alg_rec_agg = actions_alg_rec_small.groupby('recommenderId').count().reset_index()
actions_alg_rec_agg = actions_alg_rec_agg[['recommenderId','loginId']].copy()
actions_alg_rec_agg.rename(columns={"loginId":"count_rec_clicks"},inplace=True)

actions_alg_rec_agg.head()

interactions_alg_rec_percent = pd.merge(interactions_alg_group, actions_alg_rec_agg, on="recommenderId", how="left")
interactions_alg_rec_percent['ratio_action_rec_clicks'] = interactions_alg_rec_percent['count']/interactions_alg_rec_percent['count_rec_clicks']

interactions_alg_rec_percent

interactions_alg_rec_percent.to_csv('../Clean_Data/interactions_rec_ratio.csv')


### Interactions with all trailers viewed

# We need to double check this section if we need this data

interactions_alg.count()

interactions_alg.groupby("action").count()

actions_alg_total = actions_alg[actions_alg['action'].isin(["PlayedFromLaunch", "PlayedFromNext", "PlayedFromRecommendation", "PlayedFromPrevious"])]
actions_alg_total_small = actions_alg_total[['loginId','movieId','action','tstamp','userId','recommenderId']]

# Actions taken place AFTER user clicked on the recommended trailer
interactions_alg_total_agg = pd.merge(interactions_alg, actions_alg_total_small, on=['loginId','movieId'],how='inner')
interactions_alg_total_agg.drop_duplicates(["loginId","movieId","tstamp_x"],inplace=True)

interactions_alg_total_agg_small = interactions_alg_total_agg[interactions_alg_total_agg['tstamp_x']>interactions_alg_total_agg['tstamp_y']].copy()
interactions_alg_total_agg_small.count()

interactions_alg_total_agg_small.groupby("action_x").count()
# We excluded sessions with no alg assignment!

# Count the number of times an action taken place AFTER user clicked on a recommended trailer
interactions_alg_total_agg_small = interactions_alg_total_agg[interactions_alg_total_agg['tstamp_x']>interactions_alg_total_agg['tstamp_y']].copy()


interactions_alg_total_group = interactions_alg_total_agg_small.groupby(['action_x','recommenderId_x']).count().reset_index()[['action_x','recommenderId_x','loginId']]
interactions_alg_total_group.rename(columns = {"action_x":"action", "recommenderId_x":"recommenderId","loginId":"count"},inplace=True)

interactions_alg_total_group

# Action_alg_rec_small is created abooove

actions_alg_total_agg = actions_alg_total_small.groupby('recommenderId').count().reset_index()
actions_alg_total_agg = actions_alg_total_agg[['recommenderId','loginId']].copy()
actions_alg_total_agg.rename(columns={"loginId":"count_total_clicks"},inplace=True)

actions_alg_total_agg.head()

interactions_alg_total_percent = pd.merge(interactions_alg_total_group, actions_alg_total_agg, on="recommenderId", how="left")
interactions_alg_total_percent['ratio_action_total_clicks'] = interactions_alg_total_percent['count']/interactions_alg_total_percent['count_total_clicks']

interactions_alg_total_percent

interactions_alg_total_percent.to_csv('../Clean_Data/interactions_total_ratio.csv')
