==> Data_Preparation/4 - Trailers Played and (Dis)liked.ipynb <==
Data Date Range:
    <br />Start time: 2016-05-05 09:40:26
    <br />End time: 2018-01-17 01:59:42

#### This notebook generates the following table:
- rec_played_liked: a table with the recommended trailers that are played, and if the trailer is liked (by anyone at any point)
- total_played_like}d: a table with all the trailers that are played, and if the trailer is liked (by anyone at any point)

### Import and run code in another notebook

import pandas as pd
import numpy as np

%run '1 - Generate Aggregate Table with Counts Played for Algorithm.ipynb'


### Generate the dataframe with recommended trailers played and trailers liked ###

- Collected data before 2018-01-17
- Removed movielens users

# Connect to the database
import pymysql.cursors
import json
import re

connection = pymysql.connect(host="127.0.0.1",
                             user="web",
                             password="atth1132",
                             db="ML3_mirror")

try:
    with connection.cursor() as cursor:
        # This query select rows from log_trailer_actions where the action is either TrailerModalLaunched or Recommendations Refreshed
        # These two actions indicate that a recommendation list is generated. 
        
        sql_mlusers = "select userId from user_data where excludeFromExperiments='Y'"
        cursor.execute(sql_mlusers)
        mlusers = [str(item[0]) for item in cursor.fetchall()]
        mlusers = ",".join(mlusers)
        
        sql = """
        SELECT userId,loginId,movieId,isLiked,tstamp FROM trailer_likes WHERE tstamp < '2018-01-17 01:59:42' and userId NOT IN (%s) 
        """ % mlusers
        cursor.execute(sql)
        result = cursor.fetchall()
            
finally: 
    connection.close()


trailerLiked = pd.DataFrame.from_dict(list(result))
trailerLiked.columns=['userId','loginId','movieId','isLiked','tstamp']

trailerLiked

### Dataframe with recommended trailers that are clicked, and isLiked ###

# Join table of rec_played2 with table of newtrailerLiked
rec_liked = pd.merge(rec_played_alg2, trailerLiked, on=['movieId','loginId','userId'], how='left')
rec_liked.drop('Unnamed: 0',axis=1,inplace=True)

rec_liked.count()

Now, we have generated a table with the rec trailers played, and whether the trailer is liked (can be by anyone during any session)

# Write this dataframe to csv
rec_liked.to_csv("../Clean_Data/rec_played_liked.csv")


### Dataframe with all trailers viewed and isLiked ###

# Join table of rec_played2 with table of newtrailerLiked
total_trailer_info = pd.read_csv("../Clean_Data/trailers_played_info.csv")
total_liked = pd.merge(trailerLiked, total_trailer_info, on=['movieId','loginId','userId'], how='left')
# total_liked.drop('Unnamed: 0',axis=1,inplace=True)


total_liked.count()

Now, we have generated a table with total trailers played, and whether the trailer is liked (can be by anyone during any session)

total_liked.to_csv("../Clean_Data/total_played_liked.csv")
